[src] Updating working code
Update the code. It computes the rotamer that provide the maximum rotation for the first half of the given molecule.
Now the mol2 file needs to be provided as input in command whan launching the program.

+++ src/atom.cuh
@@ -6,7 +6,7 @@
 struct atom_st
 {
     unsigned int id;
-    float3 position;
+    double3 position;
 };
 
 #endif
