+++ src/rotation.cuh
@@ -9,20 +9,20 @@ class Rotation{
     
     public:
         
-        Rotation(float3 vector);
-        __host__ void getQuaternion(float4* h_quat,float4* unit_quaternion);
+        Rotation(double3 vector);
+        __host__ void getQuaternion(double4* h_quat,double4* unit_quaternion);
         //__host__ std::vector<std::vector<double>> getRotationMatrix();
         //~Quaternion();
-        atom_st* rotate(float angle, std::vector<atom_st>& atoms,float3& passingPoint);
+        atom_st* rotate(double angle, std::vector<atom_st>& atoms,double3& passingPoint);
         
 
     private:
-        float4 quaternion;
-        float4 unit_quaternion;
+        double4 quaternion;
+        double4 unit_quaternion;
         
         std::vector<std::vector<double>> rotation_matrix;
         
-        void normalize(float* quat,float* norm);//float4& vec, float *norm
+        void normalize(double* quat,double* norm);//float4& vec, float *norm
 
     };
